refactor: move style and tests files from pages directory
