+++ src/tests/404.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Custom404 from './404';
+import Custom404 from '../pages/404';
 import { RouterContext } from 'next/dist/shared/lib/router-context.shared-runtime';
 import { createMockrouter } from '../test-utils/mockRouter';
